Refactoring and package reorganization
