+++ src/main/java/io/github/doenisf/comlink4j/model/events/territorybattle/zone/recon/TerritoryReconZoneUnit.java
@@ -1,7 +1,7 @@
 package io.github.doenisf.comlink4j.model.events.territorybattle.zone.recon;
 
-import io.github.doenisf.comlink4j.model.player.unit.RelicTier;
-import io.github.doenisf.comlink4j.model.player.unit.UnitTier;
+import io.github.doenisf.comlink4j.model.unit.unit.RelicTier;
+import io.github.doenisf.comlink4j.model.unit.unit.UnitTier;
 import lombok.Getter;
 
 @Getter
