+++ src/main/java/io/github/doenisf/comlink4j/model/events/territorywar/TerritoryWarPlayerStatus.java
@@ -1,8 +1,8 @@
 package io.github.doenisf.comlink4j.model.events.territorywar;
 
 import io.github.doenisf.comlink4j.model.events.territorybattle.TerritoryPlayerUnitStatus;
-import io.github.doenisf.comlink4j.model.player.datacron.DatacronBattleStat;
-import io.github.doenisf.comlink4j.model.player.stat.UnitBattleStat;
+import io.github.doenisf.comlink4j.model.datacron.datacron.DatacronBattleStat;
+import io.github.doenisf.comlink4j.model.stat.stat.UnitBattleStat;
 import lombok.Getter;
 
 import java.util.List;
