+++ src/main/java/io/github/doenisf/comlink4j/model/events/territorywar/TerritoryWarSquad.java
@@ -1,6 +1,6 @@
 package io.github.doenisf.comlink4j.model.events.territorywar;
 
-import io.github.doenisf.comlink4j.model.player.pvp.Squad;
+import io.github.doenisf.comlink4j.model.pvp.pvp.Squad;
 import lombok.Getter;
 
 import java.util.List;
