+++ src/main/java/io/github/doenisf/comlink4j/model/events/territorywar/UnitTierSummary.java
@@ -1,7 +1,7 @@
 package io.github.doenisf.comlink4j.model.events.territorywar;
 
-import io.github.doenisf.comlink4j.model.player.unit.RelicTier;
-import io.github.doenisf.comlink4j.model.player.unit.UnitTier;
+import io.github.doenisf.comlink4j.model.unit.unit.RelicTier;
+import io.github.doenisf.comlink4j.model.unit.unit.UnitTier;
 import lombok.Getter;
 
 @Getter
